+++ packages/phrases/src/locales/pl-pl/errors/organization-invitation.ts
@@ -1,0 +1,7 @@
+const organization_invitation = {
+  invitee_already_member: 'The invitee is already a member of the organization.',
+  status_unchangeable: 'The status of the invitation cannot be changed anymore.',
+  accepted_user_email_mismatch: 'The accepted user must have the same email as the invitee.',
+};
+
+export default Object.freeze(organization_invitation);